+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Added index on `feed_data` - `feed_data_updated_at_feed_id_is_not_enriched_idx`
 - Separate firms event combination, combine firms events iteratively
+- Added disaster-ninja-02 feed (for zigzag renamed gdacs-firms feed)
+  !!! BE AWARE !!! We should change DN configuration in sync
 
 #### Changed
 
@@ -15,7 +17,7 @@
 ```yaml
 scheduler:
   feedComposition:
-    firmsAlias: firms, gdacs-firms
+    firmsAlias: firms, disaster-ninja-02
 ```
 
 #### Removed